Refactored SecretHasher, StringCipher and Password into common class Secret
