--- Application/Common/Password.cs
@@ -1,55 +1,0 @@
-﻿namespace Application.Common;
-
-using Microsoft.AspNetCore.Identity;
-
-public static class Password
-{
-    private static readonly PasswordOptions PasswordOptions = new()
-    {
-        RequiredLength = 24,
-        RequiredUniqueChars = 8,
-        RequireDigit = true,
-        RequireLowercase = true,
-        RequireNonAlphanumeric = true,
-        RequireUppercase = true
-    };
-
-    public static string Generate(PasswordOptions? opts = null)
-    {
-        opts ??= PasswordOptions;
-
-        var randomChars = new[] {
-            "ABCDEFGHJKLMNOPQRSTUVWXYZ",    // uppercase 
-            "abcdefghijkmnopqrstuvwxyz",    // lowercase
-            "0123456789",                   // digits
-            "!@$?_-"                        // non-alphanumeric
-        };
-        var rand = new Random();
-        var chars = new List<char>();
-
-        if (opts.RequireUppercase)
-            chars.Insert(rand.Next(0, chars.Count),
-                randomChars[0][rand.Next(0, randomChars[0].Length)]);
-
-        if (opts.RequireLowercase)
-            chars.Insert(rand.Next(0, chars.Count),
-                randomChars[1][rand.Next(0, randomChars[1].Length)]);
-
-        if (opts.RequireDigit)
-            chars.Insert(rand.Next(0, chars.Count),
-                randomChars[2][rand.Next(0, randomChars[2].Length)]);
-
-        if (opts.RequireNonAlphanumeric)
-            chars.Insert(rand.Next(0, chars.Count),
-                randomChars[3][rand.Next(0, randomChars[3].Length)]);
-
-        for (var i = chars.Count; i < opts.RequiredLength || chars.Distinct().Count() < opts.RequiredUniqueChars; i++)
-        {
-            var rcs = randomChars[rand.Next(0, randomChars.Length)];
-            chars.Insert(rand.Next(0, chars.Count),
-                rcs[rand.Next(0, rcs.Length)]);
-        }
-
-        return new string(chars.ToArray());
-    }
-}
